requirements.txt
dashscope
faiss-cpu
httpx
jieba
json5
jupyter>=1.0.0
langchain
langchain-community
langchain_core
langchain-experimental
modelscope>=1.9.3
moviepy
opencv-python
openpyxl
pdfminer.six
Pillow
pydantic>=2.3.0
pytest
pytest-mock
python-dotenv
requests
seaborn
sentencepiece
streamlit==1.31.0
sympy
tiktoken
transformers>=4.33
unstructured